requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.11.8
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
jsonfield==3.1.0
pipenv==2020.11.15
psycopg2==2.8.6
PyMySQL==0.10.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
virtualenv==20.2.1
virtualenv-clone==0.5.4
whitenoise==5.2.0
